==> notebooks/1-Preprocesamiento.ipynb <==
# **Preprocesamiento de datos**

## Carga de datos y librerías

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# Cargar datos
path = "C:/Users/franc/Documents/Lic. IA/Semestre 4/Machine learning" # Cambiar por la ruta de los datos
df = pd.read_csv(path + '/data/chess_games.csv')

## Exploración de datos

df.head()

# Mostrar información del dataset
df.info()

#Valores únicos por cada columna
columnas = ['Event', 'White','Black', 'Result', 'UTCDate', 'UTCTime', 'WhiteElo', 'BlackElo', 'WhiteRatingDiff', 'BlackRatingDiff', 'ECO', 'Opening', 'TimeControl', 'Termination', 'AN']
for columna in df.columns:
    valu = len(df[columna].unique())
    print(f"Valores únicos en {columna}: {valu}")

# Limpieza de datos

## Tratamiento de datos faltantes

# datos faltantes
df.isnull().sum()

Al parecer las columnas WhiteRatingDiff y BlackRatingDiff tienen el mismo número de columnas con datos faltantes, así que probablemente sean en las mismas filas, así que comprobaremos esto.

# comprobamos si las mismas filas tienen valores faltantes en las columnas BlackRatingDiff y WhiteRatingDiff
df[df['BlackRatingDiff'].isnull()].equals(df[df['WhiteRatingDiff'].isnull()])

Efectivamente, las columnas BlackRatingDiff y WhiteRatingDiff tienen valores faltantes en las mismas columnas. Al ser un dataset bastante grande, no afectará en nada el perder estos datos, por lo que serán eliminados

#borrar columnas donde hay valores faltantes en BlackRatingDiff y WhiteRatingDiff
df = df.dropna(subset=['BlackRatingDiff', 'WhiteRatingDiff'])
df.isnull().sum()

## Transformación de datos

El tipo de dato de las columnas BlackRatingDiff y WhiteRatingDiff son de tipo float pero realmente podrían ser enteros ya que no tienen números decimales, por lo que transformarán en enteros.

# Se transformamos las columnas BlackRatingDiff y WhiteRatingDiff a tipo int64
df['BlackRatingDiff'] = df['BlackRatingDiff'].astype('int64')
df['WhiteRatingDiff'] = df['WhiteRatingDiff'].astype('int64')
df.info()

#Valores únicos de la columna "Result"
print(df["Result"].value_counts())

df[df["Result"] == "*"]

Existen 3 valores con otro formato, probablemente por algún error al guardar el dato. Estos datos tienen en común que en BlackRatingDiff el valor es positivo mientras que en WhiteRatingDiff es negativo, y la partida terminó por abandono. Con los puntajes de WhiteRatingDiff y BlackRatingDiff se asume que ganaron las negras por lo que en Result, el valor debería de ser 0-1, así que se sustituirá este valor en estos registros.

#Cambiar los valores "*" por "0-1"
df.loc[df["Result"] == "*", "Result"] = "0-1"
df["Result"].value_counts()

# Valores únicos de la columna "Event"
df["Event"].value_counts()

df["Event"].unique()

Cada valor de Event se repite dos veces, una con espacios adelante y atras, y otras sin esos espacios. Se quitará dicho espacio de los valores que lo tengan.

# Eliminar espacios en blanco al inicio y al final de los valores de la columna "Event"
values_event = list(df["Event"].unique()) # Lista de valores únicos de la columna "Event"
for value in values_event: # Se itera sobre la lista de valores únicos
    df.loc[df["Event"] == value, "Event"] = value.strip() # en cada indice donde el valor sea igual al valor de la lista se elimina los espacios en blanco al inicio y al final
df["Event"].unique()

## Homogeneidad de datos

#Todos los valores de la columna "UTCDate" que tengan una longitud mayor a 10
df[df["UTCDate"].str.len() > 10]

#Todos los valores de la columna "UTCTime" que tengan una longitud mayor a 8
df[df["UTCTime"].str.len() > 8]

Se corroboró que las valores de las columnas "UTCDate" y "UTCTime" tuvieran el mismo formato.

# Se guarda el dataset limpio
df.to_csv(path + '/data/chess_games_clean.csv', index=False)

Con esto se conluye la parte del preprocesamiento. Un análisis más profundo se realizará en el notebook EDA para comprender de mejor manera los patrones y tendencias del notebook.
