==> notebooks/2-Feature_Engineering.ipynb <==
# **Feature Engineering**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

path = "C:/Users/franc/Documents/Lic. IA/Semestre 4/Machine learning" # Cambiar por la ruta de los datos
df = pd.read_csv(path + '/data/chess_games_clean.csv')
df.head()

## Tratamiento columna Result

df["Result"].value_counts()

Se cambiarán los valores de la columna result por valores numéricos de la siguiente manera:  
|Antes | Después|
|------|--------|
|'1-0' | 1 |
|'0-1' | 0 |
|'1/2-1/2' | 2 |

# Mapear los resultados a valores numéricos
result_mapping = {'1-0': 1, '0-1': 0, '1/2-1/2': 2}
df['Result'] = df['Result'].map(result_mapping) 
df["Result"].head()

## Tratamiento columna "UTCTime"

# Convertir la columna "UTCTime" a tipo datetime
df["UTCTime"] = pd.to_datetime(df["UTCTime"], format="%H:%M:%S")

# Crear columnas para la hora y el minuto
df["Hour"] = df["UTCTime"].dt.hour

# Eliminar la columna "UTCTime"
df = df.drop(columns=["UTCTime"])

# Reordenar las columnas
df = df[['Event', 'White', 'Black', 'Result', 'UTCDate', 'Hour','WhiteElo', 'BlackElo', 'WhiteRatingDiff', 'BlackRatingDiff', 'ECO', 'Opening', 'TimeControl', 'Termination', 'AN']]

df.head()

## Eliminando columna "UTCDate", "ECO", "WhiteRatingDiff" y "BlackRatingDiff"

Cómo el dataset se obtuvo sólo de las partidas jugadas en el mes de Junio, entonces no me es muy relevante el día en cuanto a que afectara a los resultados de las partidas así que se eliminará del dataset.  
También, las columnas "ECO" y "Opening" hablan ambas del tipo de apertura en la partida, pero "Opening" da más información, así que se eliminará la columna "ECO".  
Las columnas "WhiteRatingDiff" y "BlackRatingDiff" dan los puntos que se sumaron o restaron al elo de cada uno, por lo que sería un spoiler para el dataset al hacer la regresión, y cómo este ejercicio es sólo didáctico, no importará si se eliminan para entrenar el modelo,

df = df.drop(columns=["UTCDate", "ECO", "WhiteRatingDiff", "BlackRatingDiff"])
df.sample()

## Frequency Encoding para columnas con muchas categorías

columnas_cat = []
for col in df.columns:
    if df[col].dtype == object:
        if df[col].nunique() > 10:
            print(f"{col} : {df[col].nunique()}")
            columnas_cat.append(col)
print(columnas_cat)

columnas_cat.remove("AN")
print(columnas_cat)

for column in columnas_cat:
    freq_encoding = df[column].value_counts().to_dict()
    df[column] = df[column].map(freq_encoding)
df.head()

## Balanceando las clases de la columna "Result"

# Revisando el balance de clases
print(df["Result"].value_counts()/len(df) * 100)
print(df["Result"].value_counts())

Como las clases están desbalanceadas y la columna "Result" será la variable a predecir en los modelos de clasificación, se guardará un dataset con clases balanceadas. Esto también para reducir la dimensión del dataset, ya que previamente se llevaron acabo pruebas con modelos y el dataset era demasiado grande, por lo que no se entrenaba o se detenía al no soportar la dimensión del dataset.  
Cabe aclarar que este es un proyecto puramente didáctico por lo que se puede hacer esto con el dataset, ya que el objetivo es aprender a usar lo diferentes modelos de la librería scikitlearn.  
La columna "AN" donde se guardan las partidas en formato texto, se tokenizarán para probar nuevas técnicas de clasificación.

# Seleccionar solo un subconjunto de datos para el desbalance de clases
df1 = df[df["Result"] == 2]

#Tomar una muestra de 1000 datos
df2 = df[df["Result"] == 0].sample(len(df1))
df3 = df[df["Result"] == 1].sample(len(df1))
                                   
df4 = pd.concat([df1, df2, df3])

#reindexar el dataframe
df = df4.reset_index(drop=True)

## Tratamiento columna AN

La columna "AN" tiene todas las jugadas del partida en formato string. Se realizarán varios métodos de limpieza y transformación a esta columna para que pueda ser utilizable.  
Primero se eliminarán los últimos resultados ya que al final se agrega el resultado de la partida y es precisamente esto lo que queremos predecir. Este es un ejercicio didáctico y lo que se quiere es crear un modelo que prediga el resultado de la partida dependiendo de las demás variables en el dataset.   
Existen algunos comentarios entre llaves propios de la escritura de las partidas de ajedrez, estos comentarios se eliminarán de la partida.  
Después de estas transformaciónes algunos datos quedan como NAN y se remplazarán por cadenas vacías.  
Se convierte cada string en la columna "AN" en una secuencia de tokens utilizando la clase Tokenizer de Keras. Cada token generalmente representa una palabra.  
Para que todas las secuencias tengan la misma longitud (necesario para muchos modelos de aprendizaje automático), se añaden ceros al final de las secuencias más cortas hasta que todas tengan la misma longitud que la secuencia más larga.
El DataFrame de secuencias paddeadas se concatena con el DataFrame original df, añadiendo las secuencias como nuevas columnas.
Finalmente, se asegura que todos los nombres de las columnas sean strings. Esto puede ser útil para evitar errores al referenciar columnas por su nombre más adelante.

#aplicar la función anterior a toda la columna AN
df["AN"] = df["AN"].apply(lambda x: " ".join(x.split(" ")[:-1]))
df["AN"].sample()

#Eliminar el contenido que este entre llaves {} de la columna AN
df["AN"] = df["AN"].str.replace(r"\{.*\}", "", regex=True)
df["AN"].sample(20)

# Reemplazar los valores faltantes por un string vacío
df['AN'] = df['AN'].fillna('') 

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

# Tokenizar los movimientos
tokenizer = Tokenizer()
tokenizer.fit_on_texts(df['AN'])
sequences = tokenizer.texts_to_sequences(df['AN'])

# Padding de secuencias
max_sequence_length = max([len(seq) for seq in sequences])
padded_sequences = pad_sequences(sequences, maxlen=max_sequence_length)

# Convertir las secuencias paddeadas a un DataFrame
padded_sequences_df = pd.DataFrame(padded_sequences, index=df.index)

# Combinar las secuencias paddeadas con las demás columnas del DataFrame original
df = pd.concat([df, padded_sequences_df], axis=1)

df.columns = df.columns.astype(str)

## Tratamiento columnas "Termination" y "Event"

Las columnas "Termination" y "Event" se convertirán de variables categóricas a numéricas.

# label encoding
from sklearn.preprocessing import LabelEncoder
le = LabelEncoder()
df["Termination"] = le.fit_transform(df["Termination"])
df["Event"] = le.fit_transform(df["Event"])
df.head()

# guardar el dataset limpio
df.to_csv(path + '/data/chess_games_clean2.csv', index=False)
