==> notebooks/4-Regresion.ipynb <==
# Regresión

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.linear_model import Lasso
from sklearn.linear_model import Ridge
from sklearn.pipeline import Pipeline
from sklearn.neighbors import KNeighborsRegressor

path = "C:/Users/franc/Documents/Lic. IA/Semestre 4/Machine learning" # Cambiar por la ruta de los datos
df = pd.read_csv(path + '/data/chess_games_clean2.csv')
df

df = df.select_dtypes(include=['float64', 'int64'])

# Separanado las variables dependientes e independientes
X = df.drop(columns='WhiteElo')
y = df['WhiteElo']  

# Estandarizando los datos
scaler = StandardScaler()
X = scaler.fit_transform(X)

## Regresión lineal

# creando un pipeline antes de la regresión lasso
pipeline = Pipeline([('scaler', StandardScaler()), ('pca', PCA(n_components=0.90)), ('linealregression', LinearRegression())])
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
pipeline.fit(X_train, y_train)
y_pred = pipeline.predict(X_test)

#Evaluando el modelo
print('Mean squared error: %.2f' % mean_squared_error(y_test, y_pred))
print('Coefficient of determination: %.2f' % r2_score(y_test, y_pred))

## Regresión Lasso

# creando un pipeline antes de la regresión lasso
pipeline = Pipeline([('scaler', StandardScaler()), ('pca', PCA(n_components=0.90)), ('lasso', Lasso(alpha=0.1))])
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
pipeline.fit(X_train, y_train)
y_pred = pipeline.predict(X_test)


#Evaluando el modelo
print('Mean squared error: %.2f' % mean_squared_error(y_test, y_pred))
print('Coefficient of determination: %.2f' % r2_score(y_test, y_pred))

#viendo los coeficientes
df_coef = pd.DataFrame(pipeline.named_steps['lasso'].coef_, index=df.drop(columns='WhiteElo').columns, columns=['Coeficiente'])
df_coef

## Regresión Ridge

# creando un pipeline antes de la regresión lasso
pipeline = Pipeline([('scaler', StandardScaler()), ('pca', PCA(n_components=1)), ('ridge', Ridge())])
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
pipeline.fit(X_train, y_train)
y_pred = pipeline.predict(X_test)

#Evaluando el modelo
print('Mean squared error: %.2f' % mean_squared_error(y_test, y_pred))
print('Coefficient of determination: %.2f' % r2_score(y_test, y_pred))

## Selección de características

import statsmodels.api as sm
from statsmodels.api import add_constant

# Separanado las variables dependientes e independientes
X = df.drop(columns='WhiteElo')
y = df['WhiteElo']  

X_ols = X.copy()
X_ols['intercept'] = 1

model = sm.OLS(y, X_ols)
results = model.fit()

# Asumiendo que X es un DataFrame de pandas
X_ols = add_constant(X, prepend=False)
model = sm.OLS(y, X_ols)
results = model.fit()
print(results.summary())

# Obtener los p-values
pvalues = results.pvalues

# Filtrar las columnas de X_ols con p-values menores a 0.05
selected_columns = pvalues[pvalues < 0.05].index

print(selected_columns)

selected_columns = selected_columns.drop('const')
X_selected = X[selected_columns]
X_selected

scaler = StandardScaler()
X_selected = scaler.fit_transform(X_selected)

# creando un pipeline antes de la regresión lasso
pipeline = Pipeline([('scaler', StandardScaler()), ('linealregression', LinearRegression())])
X_train, X_test, y_train, y_test = train_test_split(X_selected, y, test_size=0.2, random_state=42)
pipeline.fit(X_train, y_train)
y_pred = pipeline.predict(X_test)

print('Mean squared error: %.2f' % mean_squared_error(y_test, y_pred))
print('Coefficient of determination: %.2f' % r2_score(y_test, y_pred))

plt.plot(np.arange(0, len(y_test)), np.arange(0, len(y_test)))

# Conclusión
| Regresión | R2_score |
|-----------|----------|
|LinearRegression| 0.59 |
|Lasso|0.59|
|Ridge|0.59|
|LR Selected|0.59|
  
Por el valor de los coeficientes y el comportamiento de las gráficas, podemos concluir que el Elo del jugador con piezas blancas no tiene mucha relación con cómo se desempeño la partida. Esto puede deberse a que las condiciones de cada partida son muy diferentes y ni siquiera el puntaje Elo de los jugadores de las piezas negras puede explicar del todo la variable "objetivo", ya que, si bien existe una relación entre ambos puntajes, lo cierto es que el ajedrez es un juego complejo y muy diverso, por lo que pueden darse partidas con una gran diferencia de elo entre jugadores. Una posible razón para esto, es que entre mayor sea la diferencia de puntajes, más grande es el aumento del elo para el jugador con el puntaje más bajo.  
Hablando en términos estadísticos, el elo no es una consecuencia de las demás variables, es decir, no hay una causalidad. Por ejemplo, no por que ganaron las piezas blancas y el elo de las piezas blancas era bajo, significa que eso produce que el elo del jugador con las piezas blancas se vuelva blanco, es decir, esto no es la consecuencia de las demás variables. Una de las hipótesis de la regresión lineal, es la causalidad de variable objetivo, pero este no es el caso. Esta regresión se realizo únicamente con fines didácticos y es de hecho un buen ejercicio el notar este tipo de hipótesis intrínsecas que conlleva la regresión lineal.  
En el notebook de clasificación se intentará predecir la variable "Result" de este mismo dataset, la cuál nos dice el resultado de la partida. Esta variable si demuestra la causalidad de las demás variables de este dataset, por lo que se esperan mejores resultados.
