==> notebooks/5-Clasificacion.ipynb <==
# Clasificación

#importar librería pandas y cargar el archivo csv
import pandas as pd
path = "C:/Users/franc/Documents/Lic. IA/Semestre 4/Machine learning" # Cambiar por la ruta de los datos
df = pd.read_csv(path + '/data/chess_games_clean2.csv')
df.head()

df.info()

columnas = ['Event', 'White', 'Black', 'Hour', 'WhiteElo', 'BlackElo', 'Opening','TimeControl', 'Termination']

from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import SGDClassifier
from sklearn.kernel_approximation import RBFSampler
from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.preprocessing import StandardScaler, MinMaxScaler

#lista de modelos a probar
models = {
    'MultinomialNB': MultinomialNB(),
    'RandomForestClassifier': RandomForestClassifier(max_depth=10),
    'LogisticRegression': LogisticRegression(max_iter=1000),}


## Probando modelos para datos sin "AN"

X = df[columnas]
y = df['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=90)

#preprocesamiento
scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

for model_name, model in models.items():
    model.fit(X_train, y_train)
    y_pred_test = model.predict(X_test)
    y_pred_train = model.predict(X_train)
    print(model_name)
    print("Train: ")
    print(f"Accuracy: {accuracy_score(y_train, y_pred_train)}")
    print(f"Precision: {precision_score(y_train, y_pred_train, average='weighted')}")
    print(f"Recall: {recall_score(y_train, y_pred_train, average='weighted')}")
    print(f"F1 Score: {f1_score(y_train, y_pred_train, average='weighted')}")
    print()
    print("Test: ")
    print(f"Accuracy: {accuracy_score(y_test, y_pred_test)}")
    print(f"Precision: {precision_score(y_test, y_pred_test, average='weighted')}")
    print(f"Recall: {recall_score(y_test, y_pred_test, average='weighted')}")
    print(f"F1 Score: {f1_score(y_test, y_pred_test, average='weighted')}")
    print()

## Probando modelos solo para AN

X = df.drop(columns=['Result', 'AN'])
X = X.drop(columns = columnas)
y = df['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=90)

scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

for model_name, model in models.items():
    model.fit(X_train, y_train)
    y_pred_test = model.predict(X_test)
    y_pred_train = model.predict(X_train)
    print(model_name)
    print("Train: ")
    print(f"Accuracy: {accuracy_score(y_train, y_pred_train)}")
    print(f"Precision: {precision_score(y_train, y_pred_train, average='weighted')}")
    print(f"Recall: {recall_score(y_train, y_pred_train, average='weighted')}")
    print(f"F1 Score: {f1_score(y_train, y_pred_train, average='weighted')}")
    print()
    print("Test: ")
    print(f"Accuracy: {accuracy_score(y_test, y_pred_test)}")
    print(f"Precision: {precision_score(y_test, y_pred_test, average='weighted')}")
    print(f"Recall: {recall_score(y_test, y_pred_test, average='weighted')}")
    print(f"F1 Score: {f1_score(y_test, y_pred_test, average='weighted')}")
    print()

## Probando todas las columnas

Se usan sólo los dos modelos que daban los mejores resultados en los anteriores entrenamientos

models = {
    'MultinomialNB': MultinomialNB(),
    'RandomForestClassifier': RandomForestClassifier(max_depth=10),
}

X = df.drop(columns=['Result', 'AN'])
y = df['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=90)

scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

for model_name, model in models.items():
    model.fit(X_train, y_train)
    y_pred_test = model.predict(X_test)
    y_pred_train = model.predict(X_train)
    print(model_name)
    print("Train: ")
    print(f"Accuracy: {accuracy_score(y_train, y_pred_train)}")
    print(f"Precision: {precision_score(y_train, y_pred_train, average='weighted')}")
    print(f"Recall: {recall_score(y_train, y_pred_train, average='weighted')}")
    print(f"F1 Score: {f1_score(y_train, y_pred_train, average='weighted')}")
    print()
    print("Test: ")
    print(f"Accuracy: {accuracy_score(y_test, y_pred_test)}")
    print(f"Precision: {precision_score(y_test, y_pred_test, average='weighted')}")
    print(f"Recall: {recall_score(y_test, y_pred_test, average='weighted')}")
    print(f"F1 Score: {f1_score(y_test, y_pred_test, average='weighted')}")
    print()

## Gridsearch para RandomForest (no se terminó de entrenar)

X = df.drop(columns=['Result', 'AN'])
y = df['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=90)

scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

# GridSearchCV para RandomForestClassifier
param_grid = {
    'n_estimators': [10, 100],
    'max_depth': [10, 100],
}

model = RandomForestClassifier()
grid_search = GridSearchCV(model, param_grid)
grid_search.fit(X_train, y_train)
print(grid_search.best_params_)
y_pred = grid_search.predict(X_test)

print(f"Accuracy: {accuracy_score(y_test, y_pred)}")
print(f"Precision: {precision_score(y_test, y_pred, average='weighted')}")
print(f"Recall: {recall_score(y_test, y_pred, average='weighted')}")
print(f"F1 Score: {f1_score(y_test, y_pred, average='weighted')}")

## Resultados
|Modelo| sin AN|con AN| todas|
|------|-------|------|------|
|MultinomialNB| 0.39|0.65|0.67|
|RandomForest|0.48|0.70| 0.72|
|Logistic| 0.46|0.68| |
|RRN|0.33|0.36| |
|LSTM| | | 0.73|

# Extra: usando Red neuronal LSTM

### Red neuronal LSTM

df1 = df.sample(6000)

X = df1.drop(columns=['Result', 'AN'])
X = X.drop(columns = columnas)
y = df1['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=90)


from keras.models import Sequential
from keras.layers import Embedding, LSTM, Dense
from keras.optimizers import Adam

# Crear el modelo LSTM
model = Sequential()
model.add(Embedding(input_dim= X.max().max() + 1, output_dim=128, input_length=1))
model.add(LSTM(units=128))
model.add(Dense(units=4, activation='sigmoid'))

# Compilar el modelo
optimizer = Adam(learning_rate=0.01) 
model.compile(optimizer=optimizer, loss='sparse_categorical_crossentropy', metrics=['accuracy'])

# Entrenar el modelo
model.fit(X_train, y_train, epochs=2, batch_size=64, validation_data=(X_test, y_test))

# Evaluar el modelo
loss, accuracy = model.evaluate(X_test, y_test)
print(f'Accuracy: {accuracy}')


En algún momento sí se entrenaba relativamente rápido pero después cada época tardaba en entrenarse 2 horas. El siguiente resultado es de un modelo que se entrenó en media hora

import numpy as np
from sklearn.metrics import confusion_matrix
#matriz de confusion y metricas
y_pred = model.predict(X_test)
y_pred = [np.argmax(pred) for pred in y_pred]
confusion_matrix(y_test, y_pred)

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
print(f"Accuracy: {accuracy_score(y_test, y_pred)}")
print(f"Precision: {precision_score(y_test, y_pred, average='weighted')}")
print(f"Recall: {recall_score(y_test, y_pred, average='weighted')}")
print(f"F1 Score: {f1_score(y_test, y_pred, average='weighted')}")

# Conclusión
Podemos notar en las métricas de evaluación que es difícil para un modelo predecir, a partir de las características de las partidas, cómo resultará el resultado de la partida. Esto debido a que en partidas en línea, los usuarios son demasiado diferentes en cuanto a experiencia, por lo que hay todo tipo de partidas. Otro factor es el tiempo, ya que algunas partidas terminan por el tiempo así que puede terminar abruptamente una partida y no tener un patrón al momento de entrenar el modelo. Otras partidas en cambio pueden durar demasiado tiempo y no ser realmente una partida compleja, por lo que el modelo tampoco puede encontrar ciertos comportamientos. Quizá si se tuviera el dato de en qué tiempo se hace cada jugada, esto podría mejorar el modelo.  
Aún cuando se probó una red neuronal LSTM y se realizó una tokenización y un padding de secuencias, el accuracy no fue del todo bueno. Mi interpretación sobre esto es que no necesariamente existe una interpretación sintáctica como tal en las jugadas de cada partida, por lo que no necesariamente el modelo puede encontrar patrones para el entrenamiento. Quizás otra representación para las partidas podría resultar en un mejor resultado para los modelos.
